backend/requirements.txt
# Trade Mart Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
